backend/requirements.txt
openai-whisper
torch
numpy
torchaudio
soundfile
fastapi
uvicorn
python-multipart
ffmpeg
httpx
mlx-whisper
